added src and test directory and created download_temperature_data.py & test_read_data.py in src and test
