+++ test/test_read_data.py
@@ -1,0 +1,6 @@
+import pytest
+import sys
+import os
+import pandas as pd
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+from src.write_csv import write_csv
